fixed bug for inserting
